+++ vms-ide/src/python/debugConfig.ts
@@ -11,10 +11,14 @@ import { DebugConfiguration, WorkspaceFolder } from "vscode";
 export interface IPythonLaunchRequestArguments extends DebugProtocol.LaunchRequestArguments {
     /** a script to execute. */
     script: string;
+    /** a VMS command to execute before. */
+    pre_launch?: string;
     /** a port to debug. */
     port?: string;
     /** command line arguments */
     arguments?: string;
+    /** python arguments */
+    python_args?: string;
     /** current workspace folder */
     workspace: WorkspaceFolder;
 }
@@ -22,10 +26,14 @@ export interface IPythonLaunchRequestArguments extends DebugProtocol.LaunchReque
 export interface IPythonDebugConfiguration extends DebugConfiguration {
     /** a script to execute. */
     script: string;
+    /** a VMS command to execute before. */
+    pre_launch?: string;
     /** a port to debug. */
     port?: string;
     /** command line arguments */
     arguments?: string;
+    /** python arguments */
+    python_args?: string;
     /** current workspace folder */
     workspace: WorkspaceFolder;
 }
@@ -35,5 +43,9 @@ export function isPythonDebugConfiguration(candidate: any): candidate is IPython
         typeof candidate.type === "string" &&
         typeof candidate.name === "string" &&
         typeof candidate.request === "string" &&
+        (typeof candidate.python_args === undefined || typeof candidate.python_args === "string") &&
+        (typeof candidate.arguments === undefined || typeof candidate.arguments === "string") &&
+        (typeof candidate.pre_launch === undefined || typeof candidate.pre_launch === "string") &&
+        (typeof candidate.port === undefined || typeof candidate.port === "number") &&
         typeof candidate.script === "string";
 }