+++ vms-ide/src/vms_debug/net/shell-session.ts
@@ -235,7 +235,7 @@ export class ShellSession
 
                     if (this.disconnect)
                     {
-                        this.DisconectSession(true, localize("proram.completed", ": The program completed."));//close SSH session
+                        this.DisconectSession(true, localize("proram.completed", ": The program is completed."));//close SSH session
                         this.readyCmd = true;
                         this.extensionDataCb(this.mode, TypeDataMessage.typeData, this.resultData);
                         this.resultData = "";
@@ -360,7 +360,7 @@ export class ShellSession
 
         if (this.logFn)
         {
-            this.logFn(LogType.debug, () => localize("conn.closed", "Connection closed."));
+            this.logFn(LogType.debug, () => localize("conn.closed", "Connection is closed."));
         }
 
         if(code)